+++ src/main.js
@@ -49,6 +49,3 @@ Hooks.on('canvasPan', () => clearConditionals())
 Hooks.on('renderChatMessage', renderChatMessage)
 
 Hooks.on('renderSceneConfig', renderSceneConfig)
-
-// TODO sneak
-// TODO add concealed setting in scene config