+++ CHANGELOG.md
@@ -1,1 +1,7 @@
-# 1.0.0
+todo
+
+-   sneak ?
+-   add concealed setting in scene config
+-   hide combatant from tracker is unnoticed
+
+# 0.6.0